Web_Automation_Framework/requirements.txt
selenium==4.15.0
requests==2.31.0
behave==1.2.6
python-dotenv==1.0.0
webdriver-manager==4.0.1
pytest==7.4.3
allure-behave==2.13.2
beautifulsoup4==4.12.2
jsonschema==4.19.2
urllib3==2.0.7
faker==19.12.0 